Initial Prototype commit containing
Initial screens for:

- New Classroom
- Add Student Dialog
- Add Course Dialog
- Initial layout

As backend is not yet complete the code uses
a set of hard-coded global data in the services
classes

Some lint checks disabled and prop-types not yet configured.
Basic tests exist

For this state of the prototype only LMS is needed.

All data is temporary. when page is refresh the data is reset.

Redux layer is at a basic level.

Ready for a basic demo though.