Some intermediate changes:
modified temporarily the fetch of all courses from the system. At the moment it only returns
an array of course run ids. Hopefully this will be expanded with more course information
such as imageURLs, description etc. as will save on API calls.

Removed the AddCourse dispatch as this will no be needed at present.

Add code to add a course to the classroom - but this is untested as Discovery/Classroom is
giving me (Salik) huge issues with data consistency.