Issue #6;
The SVG seems to want to maintain it's proportions so when I set a height on to the image there are gaps left at top and bottom of the area. Maybe the SVG just needs to be made taller?

I have adjusted the text position and size to be like the Figma.